requirements.txt
BigQuery-Python==0.1.1
Pillow==2.7.0
beautifulsoup4==4.3.2
cffi==0.8.6
cryptography==0.7.1
enum34==1.0.4
google-api-python-client==1.3.1
httplib2==0.9
matplotlib==1.4.2
mock==1.0.1
nose==1.3.4
numpy==1.9.1
oauth2client==1.4.5
oauthlib==0.7.2
pandas==0.15.2
patsy==0.3.0
plotdevice==0.9.4
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
rsa==3.1.4
scipy==0.14.1
simplejson==3.6.5
six==1.9.0
statsmodels==0.6.1
twython==3.2.0
uritemplate==0.6
wsgiref==0.1.2
